# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /usuarios:
    x-swagger-router-controller: usuarios.controller
    get:
      description: Mostrar los usuarios en lista (get the game system list)
      operationId: getUsuarios
      #parameters:
      #  - name: name
      #    in: query
      #    type: string
      #    required: false
      #  - name: sort
      #    in: query
      #   type: string
      #    required: false
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetUsuariosListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Agrega un nuevo usuario a la lista (add a new game system to the list)
      operationId: createUsuarios
      parameters:
        - in: body
          name: Usuarios (gamesystem)
          description: El usuario a crear (The Game System to create.)
          schema:
            $ref: "#/definitions/Usuarios"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GetUsuariosResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  
  /usuarios/{id}:
    x-swagger-router-controller: usuarios.controller
    get:
      description: Dame un usuario por id (get a game system)
      operationId: getUsuariosByPk
      parameters:
        - name: id
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetUsuariosResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Actualizar los usuarios por id (update a game system)
      operationId: updateUsuarios
      # define the parameters
      parameters:
        - name: id
          description: Usuario Id (game system id)
          type: integer
          in: path
          required: true
        - in: body
          name: Usuario (gamesystem)
          description: El usuario a actualizar (The Game System to update.)
          schema:
            $ref: "#/definitions/Usuarios"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetUsuariosResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Borrar unusuario por id (delete a game system)
      operationId: deleteUsuarios
      # define the parameters
      parameters:
        - name: id
          description: Usuario id (game system id)
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  Usuarios:
    type: object
    properties:
      #id_usuario:
      #  type: integer
      #  description: Id del usuario (Id of the Game System)
      nombreUsuario:
        type: string
        description: Nombre del usuario (Name of the Game System)
      apellido:
        type: string
        description: Apellido del usuario (Description of the zone)
      #image:
      #  type: string
      #  format: binary
      email:
        type: string
        description: Email del usuario
      password:
        type: string
        description: Password del usuario
      fechaNacimiento:
        type: string
        format: date
        description: Fecha de Nacimiento del usuario
      telefono:
        type: string
        description: Telefono del usuario
    required:
      #- id_usuario
      - nombreUsuario
      - apellido
      - email
      - password
      - fechaNacimiento
      - telefono
      
  GetUsuariosResponse:
    type: object
    properties:
      id_usuario:
        type: integer
        description: Id del usuario (Id of the Game System)
      nombreUsuario:
        type: string
        description: Nombre del usuario (Name of the Game System)
      apellido:
        type: string
        description: Apellido del usuario (Description of the zone)
      image:
        type: string
        format: binary
      email:
        type: string
        description: Email del usuario
      password:
        type: string
        description: Password del usuario
      fechaNacimiento:
        type: integer
        description: Fecha de Nacimiento del usuario
      telefono:
        type: string
        description: Telefono del usuario
  GetUsuariosListResponse:
    required:
      - gamesystems
    properties:
      gamesystems:
        type: array
        items:
          $ref: "#/definitions/GetUsuariosResponse"
  GeneralResponse:
    type: object
    properties:
      success:
        type: integer
        description: returns 1 if successful
      description:
        type: string
        description: a short comment
    required:
      - success
      - description
  ErrorResponse:
    required:
      - message
    properties:
      code:
        type: number
      message:
        type: string
      description:
        type: string